server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "python-socketio>=5.13.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipdb>=0.13.13",
    "isort>=6.0.1",
]
